backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
python-multipart==0.0.9
pypdf==4.3.1
